Separate and add Page/function specific JS and CSS to separate file
All of our jQuery and most of our CSS is bunched together into one file. This makes identifying functionality of the JS or CSS more difficult.
Ideally, if there is a new or temporary code that is specific to a particular page (v.s., global functionality) it is often advantageous to contain that code in one file.

In this case, there is a new front page function that will most likely be retired as some point.

Signed-off-by: Thomas Wagner <twagner@ivlab.com>