Fix jQuery error on Admin pages
Admin functions in DS, Views and other places were failing (little spinny thing spinning forever). Turns out that the conflict comes from Owl Carousel. We were loading it on all pages, which is not only unnecessary  and inefficient - it was causing the error.

Moved the calls to Owl to a separate JS file that is only loaded on front page loads